refactor: rename category and product classes, update Flutter root paths, and clean up unused files
