--- frontend/e_commerce/.dart_tool/extension_discovery/vs_code.json
@@ -1,1 +1,0 @@
-{"version":2,"entries":[{"package":"e_commerce","rootUri":"../","packageUri":"lib/"}]}
