+++ frontend/e_commerce/.dart_tool/package_config.json
@@ -344,7 +344,7 @@
       "languageVersion": "3.4"
     }
   ],
-  "generated": "2025-01-11T12:05:38.151122Z",
+  "generated": "2025-01-12T11:55:44.389753Z",
   "generator": "pub",
   "generatorVersion": "3.4.1",
   "flutterRoot": "file:///E:/flutter",
