+++ frontend/e_commerce/lib/src/layout/layout_controller.dart
@@ -1,6 +1,6 @@
 
 import 'package:e_commerce/src/Favorites/favorite_view.dart';
-import 'package:e_commerce/src/category/category_view.dart';
+import 'package:e_commerce/src/categories/categories_view.dart';
 import 'package:e_commerce/src/products/products_view.dart';
 import 'package:e_commerce/src/setting/setting_view.dart';
 import 'package:get/get.dart';
@@ -9,7 +9,7 @@ class LayoutController extends GetxController {
   RxInt currentIndex = 0.obs;
 
   final tabs =  [
-    const ProductsSecreen(),
+    const ProductsView(),
   const  CategoryScreen(),
    const FavoritesScreen(),
    const SettingScreen(),
