+++ frontend/e_commerce/lib/src/products/products_view.dart
@@ -6,8 +6,8 @@ import 'package:e_commerce/theme/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-class ProductsSecreen extends StatelessWidget {
-  const ProductsSecreen({super.key});
+class ProductsView extends StatelessWidget {
+  const ProductsView({super.key});
 
   @override
   Widget build(BuildContext context) {
