+++ frontend/e_commerce/lib/src/products_category/products_view.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class ProductsCategoryView extends StatelessWidget {
+
+
+  const ProductsCategoryView() ;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Saeed'),
+      ),
+      body: 
+      Container()
+    );
+  }
+}
+
