+++ frontend/e_commerce/macos/Flutter/ephemeral/Flutter-Generated.xcconfig
@@ -1,5 +1,5 @@
 // This is a generated file; do not edit or check into version control.
-FLUTTER_ROOT=C:\src\flutter
+FLUTTER_ROOT=E:\flutter
 FLUTTER_APPLICATION_PATH=E:\backend-fronted-sales-management\frontend\e_commerce
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_BUILD_DIR=build
