+++ frontend/e_commerce/macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=C:\src\flutter"
+export "FLUTTER_ROOT=E:\flutter"
 export "FLUTTER_APPLICATION_PATH=E:\backend-fronted-sales-management\frontend\e_commerce"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
